+++ scripts/templates/readme.js
@@ -35,6 +35,36 @@ Alternatively, the same solutions could be imported via SCSS!
 @import "~@fontsource/<%= fontId %>/300-italic.css";
 \`\`\`
 
+For more advanced setups, you can use our highly customisable Sass mixins that can modify many of the existing @font-face variables.
+
+\`\`\`scss
+@import "~@fontsource/<%= fontId %>/scss/mixins";
+
+<% if (type == "google") { %>// Uses a unicode-range map to automatically generate multiple @font-face rules.
+@include fontFace(
+  $weight: 500,
+  $display: fallback,
+  $fontDir: "~@fontsource/<%= fontId %>/files"
+);
+
+// Fully customisable single @font-face mixin.
+@include fontFaceCustom(
+  $weight: 600,
+  $display: optional,
+  $woff2Path: "~@fontsource/<%= fontId %>/files/<%= fontId %>-latin-500-normal.woff2",
+  $unicodeRange: false
+);<% } else { %>// Fully customisable single @font-face mixin.
+@include fontFaceCustom(
+  $weight: 600,
+  $display: optional,
+  $woff2Path: "~@fontsource/<%= fontId %>/files/<%= fontId %>-latin-500-normal.woff2",
+  $unicodeRange: false
+);<% } %>
+// More options available in link below.
+\`\`\`
+
+You can see all of the existing inputtable mixin variables [here](https://github.com/fontsource/fontsource/tree/master/packages/<%= fontId %>/scss/mixins.scss).
+
 _These examples may not reflect actual compatibility. Please refer below._
 
 Supported variables: